Update wccntCNTbuild.tcl
A summmary:

1.- 
Now, "wccnt CNTbuild" can create CNTs up to 1 um length.  Before, it only worked up to 80 nm,  due to a limitation the number of atoms within a resid.  Larger CNTs (> 1um) are (still) limited by the PDB format.


2.- 
There is a new flag "-shiftXYZ", to add a vector displacement; that is, to move the CNT.


3.- 
The flag "-resID" has been removed.  The resids are assigned automatically now.


4.-
The flag "-resName" must be a string of 3 characters.  Compulsory.